QPID-7794: [Java Broker] Periodically report the number of bytes evacuated from memory by flow to disk
These are the Store parts required.

+++ broker-core/src/main/java/org/apache/qpid/server/queue/AbstractQueue.java
@@ -3919,7 +3919,7 @@ public abstract class AbstractQueue<X extends AbstractQueue<X>>
 
         void reportFlowToDiskStatusIfNecessary(final long estimatedQueueSize, final long targetQueueSize)
         {
-            final int allocatedDirectMemorySize = QpidByteBuffer.getAllocatedDirectMemorySize();
+            final long allocatedDirectMemorySize = QpidByteBuffer.getAllocatedDirectMemorySize();
             if (estimatedQueueSize > targetQueueSize
                 || allocatedDirectMemorySize > _flowToDiskThreshold)
             {
